Feat: status 체인지를 taskHandler.ts로 변경
